Add files via upload
Code along project from Youtube (https://www.youtube.com/watch?v=NgsHgBVPXpM&t=1938s ) . This was by far the hardest code along I have done as of today. He did not talk the entire video and a lot of stuff was set up outside of this video so I had to do a lot of work on my own. The HTML is mostly the same but I changed the colors and imagery to match a mock website called Papicito. This is a cafe we always joke with my dad about starting one day when he retires.